Chore: run playwright tests in parallel
