+++ playwright.config.js
@@ -4,22 +4,21 @@ import { defineConfig, devices } from "@playwright/test";
  * @see https://playwright.dev/docs/test-configuration
  */
 export default defineConfig({
-	testDir: "./tests",
-	fullyParallel: false,
-	forbidOnly: !!process.env.CI,
-	retries: process.env.CI ? 2 : 0,
-	timeout: 15000, // 15s (default 30s)
-	expect: { timeout: 2500 }, // 2.5s (default 5s)
-	workers: 1, // run testfiles serially to avoid port and database conflicts
-	reporter: "html",
-	use: {
-		baseURL: "http://127.0.0.1:7070",
-		trace: "on-first-retry",
-	},
-	projects: [
-		{
-			name: "chromium",
-			use: { ...devices["Desktop Chrome"] },
-		},
-	],
+  testDir: "./tests",
+  forbidOnly: !!process.env.CI,
+  retries: process.env.CI ? 2 : 0,
+  timeout: 15000, // 15s (default 30s)
+  expect: { timeout: 2500 }, // 2.5s (default 5s)
+  workers: process.env.CI ? 3 : 8,
+  reporter: "html",
+  use: {
+    baseURL: "http://127.0.0.1:7070",
+    trace: "on-first-retry",
+  },
+  projects: [
+    {
+      name: "chromium",
+      use: { ...devices["Desktop Chrome"] },
+    },
+  ],
 });
