+++ tests/config-meters.spec.js
@@ -1,0 +1,85 @@
+import { test, expect } from "@playwright/test";
+import { start, stop, restart, baseUrl } from "./evcc";
+import { startSimulator, stopSimulator, simulatorUrl, simulatorHost } from "./simulator";
+
+const CONFIG_EMPTY = "config-empty.evcc.yaml";
+
+test.use({ baseURL: baseUrl() });
+
+test.beforeAll(async () => {
+  await start(CONFIG_EMPTY, "password.sql");
+  await startSimulator();
+});
+test.afterAll(async () => {
+  await stop();
+  await stopSimulator();
+});
+
+async function login(page) {
+  await page.locator("#loginPassword").fill("secret");
+  await page.getByRole("button", { name: "Login" }).click();
+}
+
+async function enableExperimental(page) {
+  await page
+    .getByTestId("generalconfig-experimental")
+    .getByRole("link", { name: "change" })
+    .click();
+  await page.getByLabel("Experimental 🧪").click();
+  await page.getByRole("button", { name: "Close" }).click();
+}
+
+test.describe("meters", async () => {
+  test("create, edit and remove battery meter", async ({ page }) => {
+    // setup test data for mock openems api
+    await page.goto(simulatorUrl());
+    await page.getByLabel("Battery Power").fill("-2500");
+    await page.getByLabel("Battery SoC").fill("75");
+    await page.getByRole("button", { name: "Apply changes" }).click();
+
+    await page.goto("/#/config");
+    await login(page);
+    await enableExperimental(page);
+
+    await expect(page.getByTestId("battery")).toHaveCount(0);
+
+    // create #1
+    await page.getByRole("button", { name: "Add solar or battery" }).click();
+
+    const meterModal = page.getByTestId("meter-modal");
+    await meterModal.getByRole("button", { name: "Add battery meter" }).click();
+    await meterModal.getByLabel("Manufacturer").selectOption("OpenEMS");
+    await meterModal.getByLabel("IP address or hostname").fill(simulatorHost());
+    await expect(meterModal.getByRole("button", { name: "Validate & save" })).toBeVisible();
+    await meterModal.getByRole("link", { name: "validate" }).click();
+    await expect(meterModal.getByText("SoC: 75.0%")).toBeVisible();
+    await expect(meterModal.getByText("Power: -2.5 kW")).toBeVisible();
+    await meterModal.getByRole("button", { name: "Save" }).click();
+    await expect(page.getByTestId("battery")).toBeVisible(1);
+    await expect(page.getByTestId("battery")).toContainText("openems");
+
+    // edit #1
+    await page.getByTestId("battery").getByRole("button", { name: "edit" }).click();
+    await meterModal.getByLabel("Battery capacity in kWh").fill("20");
+    await meterModal.getByRole("button", { name: "Validate & save" }).click();
+
+    await expect(page.getByTestId("battery")).toBeVisible(1);
+    await expect(page.getByTestId("battery")).toContainText("openems");
+    await expect(page.getByTestId("battery").getByText("SoC: 75.0%")).toBeVisible();
+    await expect(page.getByTestId("battery").getByText("Power: -2.5 kW")).toBeVisible();
+    await expect(page.getByTestId("battery").getByText("Capacity: 20.0 kWh")).toBeVisible();
+
+    // restart and check in main ui
+    await restart(CONFIG_EMPTY);
+    await page.goto("/");
+    await page.getByTestId("visualization").click();
+    await expect(page.getByTestId("energyflow")).toContainText("Battery charging75%2.5 kW");
+
+    // delete #1
+    await page.goto("/#/config");
+    await page.getByTestId("battery").getByRole("button", { name: "edit" }).click();
+    await meterModal.getByRole("button", { name: "Delete" }).click();
+
+    await expect(page.getByTestId("battery")).toHaveCount(0);
+  });
+});
