+++ tests/evcc.js
@@ -2,12 +2,24 @@ import fs from "fs";
 import waitOn from "wait-on";
 import axios from "axios";
 import { exec, execSync } from "child_process";
-import playwrightConfig from "../playwright.config.js";
+import os from "os";
+import path from "path";
 
-const BASE_URL = playwrightConfig.use.baseURL;
-
-const DB_PATH = "./evcc.db";
 const BINARY = "./evcc";
+
+function port() {
+  const index = process.env.TEST_WORKER_INDEX * 1;
+  return 11000 + index;
+}
+
+export function baseUrl() {
+  return `http://localhost:${port()}`;
+}
+
+function dbPath() {
+  const file = `evcc-${port()}.db`;
+  return path.join(os.tmpdir(), file);
+}
 
 export async function start(config, sqlDumps) {
   await _clean();
@@ -39,14 +51,17 @@ export async function cleanRestart(config, sqlDumps) {
 async function _restoreDatabase(sqlDumps) {
   const dumps = Array.isArray(sqlDumps) ? sqlDumps : [sqlDumps];
   for (const dump of dumps) {
-    console.log("loading database", dump);
-    execSync(`sqlite3 ${DB_PATH} < tests/${dump}`);
+    console.log("loading database", dbPath(), dump);
+    execSync(`sqlite3 ${dbPath()} < tests/${dump}`);
   }
 }
 
 async function _start(config) {
+  const configFile = config.includes("/") ? config : `tests/${config}`;
   console.log("starting evcc", { config });
-  const instance = exec(`EVCC_DATABASE_DSN=${DB_PATH} ${BINARY} --config tests/${config}`);
+  const instance = exec(
+    `EVCC_NETWORK_PORT=${port()} EVCC_DATABASE_DSN=${dbPath()} ${BINARY} --config ${configFile}`
+  );
   instance.stdout.pipe(process.stdout);
   instance.stderr.pipe(process.stderr);
   instance.on("exit", (code) => {
@@ -54,20 +69,21 @@ async function _start(config) {
       throw new Error("evcc terminated", code);
     }
   });
-  await waitOn({ resources: [BASE_URL] });
+  await waitOn({ resources: [baseUrl()] });
 }
 
 async function _stop() {
   console.log("shutting down evcc");
-  const res = await axios.post(BASE_URL + "/api/auth/login", { password: "secret" });
+  const res = await axios.post(`${baseUrl()}/api/auth/login`, { password: "secret" });
   console.log(res.status, res.statusText);
   const cookie = res.headers["set-cookie"];
-  await axios.post(BASE_URL + "/api/system/shutdown", {}, { headers: { cookie } });
+  await axios.post(`${baseUrl()}/api/system/shutdown`, {}, { headers: { cookie } });
 }
 
 async function _clean() {
-  if (fs.existsSync(DB_PATH)) {
-    console.log("delete database", DB_PATH);
-    fs.unlinkSync(DB_PATH);
+  const db = dbPath();
+  if (fs.existsSync(db)) {
+    console.log("delete database", db);
+    fs.unlinkSync(db);
   }
 }
