+++ tests/limits.spec.js
@@ -1,8 +1,10 @@
 import { test, expect } from "@playwright/test";
-import { start, stop } from "./evcc";
+import { start, stop, baseUrl } from "./evcc";
 import { startSimulator, stopSimulator, SIMULATOR_URL } from "./simulator";
 
 const CONFIG = "simulator.evcc.yaml";
+
+test.use({ baseURL: baseUrl() });
 
 test.beforeAll(async () => {
   await startSimulator();
