+++ tests/plan.spec.js
@@ -1,5 +1,7 @@
 import { test, expect } from "@playwright/test";
-import { start, stop } from "./evcc";
+import { start, stop, baseUrl } from "./evcc";
+
+test.use({ baseURL: baseUrl() });
 
 const CONFIG = "plan.evcc.yaml";
