+++ tests/smart-cost.spec.js
@@ -1,6 +1,8 @@
 import { test, expect } from "@playwright/test";
-import { start, stop } from "./evcc";
+import { start, stop, baseUrl } from "./evcc";
 import { startSimulator, stopSimulator, SIMULATOR_URL } from "./simulator";
+
+test.use({ baseURL: baseUrl() });
 
 const CONFIG = "simulator.evcc.yaml";
